fix: separate services as a standalone module
