+++ src/scripts/modules/services/ActionCreators.js
@@ -1,8 +1,8 @@
 import dispatcher from '../../Dispatcher';
-import constants from './ServicesConstants';
+import constants from './Constants';
 
 module.exports = {
-  receiveAllServices: function(servicesRaw) {
+  receive: function(servicesRaw) {
     return dispatcher.handleViewAction({
       type: constants.ActionTypes.SERVICES_LOAD_SUCCESS,
       services: servicesRaw
